Take user groups into account when rendering menu
When a category access is restricted by user group, the category is still displayed in the menu, even if the user is not a member of one of those group. It is described here : https://github.com/PrestaShop/PrestaShop/issues/9924
There has been 2 different corrections that have been proposed on this thread, but none of them has been merged in the module (there is some commits for at least one proposition in the dev branch, but not in master branch).
I think my changes are a bit simpler. I use the "user_groups" property that was already defined in the Ps_MainMenu class, but never initialized.